move sensor information to new sensors.py file
